// be/hyper-express/src/constants/types.ts
import {Types} from "mongoose";
//import {Request} from "express";
import {Request} from "hyper-express";
export enum UserRole {
  Regular = 'Regular',
  Admin = 'Admin'
}

export enum PostType {
  Comment = 'C',
  Story = 'S',
  TimeLine = 'TL'
}

export enum PostReactType {
  Like = 'like',
  Dislike = 'dislike',
  Love = 'love',
  Care = 'care',
  Haha = 'haha',
  Wow = 'wow',
  Sad = 'sad',
  Angry = 'angry',
  UnReact = ''
}

export interface AuthUser {
  _id: Types.ObjectId,
  email: string,
  password: string,
  role: UserRole
}

export interface AuthRequest extends Request {
  user: AuthUser
}